requirements.txt
aiohttp==3.5.4
async-timeout==3.0.1
attrs==19.1.0
awscli==1.16.100
awsebcli==3.14.10
blessed==1.15.0
botocore==1.12.90
cached-property==1.5.1
cement==2.8.2
certifi==2019.3.9
chardet==3.0.4
colorama==0.3.9
discord.py==1.0.1
docker==3.7.0
docker-compose==1.23.2
docker-pycreds==0.4.0
dockerpty==0.4.1
docopt==0.6.2
docutils==0.14
idna==2.7
jmespath==0.9.3
multidict==4.5.2
pathspec==0.5.5
pyasn1==0.4.5
pycodestyle==2.5.0
rsa==3.4.2
s3transfer==0.2.0
semantic-version==2.5.0
six==1.11.0
termcolor==1.1.0
texttable==0.9.1
uvloop==0.12.2
websocket-client==0.54.0
websockets==6.0
yarl==1.3.0
